% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{cf}
\alias{cf}
\title{Covariance function}
\usage{
cf(x, y, sigmaf = 1, l = 1, sigman = 0)
}
\arguments{
\item{x}{First value}

\item{y}{Second value}

\item{sigmaf}{See equation above}

\item{l}{See equation above}

\item{sigman}{See equation above}
}
\value{
Calculated covariance
}
\description{
Covariance function
}
\details{
Covariance function for - The squared exponential covariance function
\eqn{k_{x,y} = \sigma_f^2 exp(-frac{1}{2l^2}(x - y)^2) + \delta_{xy} \sigma_n^2}
}
\author{
Wajid Jawaid
}
