% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{cohens_d}
\alias{cohens_d}
\title{Compute Cohen's d from means and standard deviations}
\usage{
cohens_d(m1, m2, sd1, sd2 = NULL, n1 = NULL, n2 = NULL)
}
\arguments{
\item{m1}{Mean of group 1}

\item{m2}{Mean of group 2}

\item{sd1}{Standard deviation of group 1}

\item{sd2}{Standard deviation of group 2 (if NULL uses sd1)}

\item{n1}{Sample size of group 1 (for pooled SD)}

\item{n2}{Sample size of group 2 (for pooled SD)}
}
\value{
Cohen's d
}
\description{
Compute Cohen's d from means and standard deviations
}
\examples{
cohens_d(m1 = 10, m2 = 12, sd1 = 4, sd2 = 4)
}
