% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{cohens_f}
\alias{cohens_f}
\title{Compute Cohen's f from group means and SD}
\usage{
cohens_f(means, sd_within, n = NULL)
}
\arguments{
\item{means}{Vector of group means}

\item{sd_within}{Within-group standard deviation}

\item{n}{Vector of group sample sizes (or single n for equal groups)}
}
\value{
Cohen's f
}
\description{
Compute Cohen's f from group means and SD
}
\examples{
cohens_f(means = c(10, 12, 14), sd_within = 8)
}
