% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{cohens_h}
\alias{cohens_h}
\title{Compute Cohen's h from two proportions}
\usage{
cohens_h(p1, p2)
}
\arguments{
\item{p1}{First proportion}

\item{p2}{Second proportion}
}
\value{
Cohen's h
}
\description{
Compute Cohen's h from two proportions
}
\examples{
cohens_h(0.50, 0.70)
}
