% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{cohens_w}
\alias{cohens_w}
\title{Compute Cohen's w from observed and expected proportions}
\usage{
cohens_w(p_obs, p_exp)
}
\arguments{
\item{p_obs}{Vector of observed proportions}

\item{p_exp}{Vector of expected proportions under H0}
}
\value{
Cohen's w
}
\description{
Compute Cohen's w from observed and expected proportions
}
\examples{
cohens_w(p_obs = c(0.40, 0.35, 0.25), p_exp = c(0.33, 0.33, 0.34))
}
