% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{covMat}
\alias{covMat}
\title{Calculate Covariance matrix}
\usage{
covMat(x, y, cvFunc)
}
\arguments{
\item{x}{Rows of covariance matrix}

\item{y}{Columns of covariance matrix}

\item{cvFunc}{Covariance function to be used see \code{\link{cf}}}
}
\value{
Returns a covariance matrix
}
\description{
Calculate Covariance matrix
}
\details{
Calculate Covariance matrix
}
\author{
Wajid Jawaid
}
