% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{d_to_or}
\alias{d_to_or}
\title{Convert Cohen's d to odds ratio}
\usage{
d_to_or(d)
}
\arguments{
\item{d}{Cohen's d}
}
\value{
Odds ratio
}
\description{
Convert Cohen's d to odds ratio
}
\examples{
d_to_or(0.5)
}
