% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{d_to_r}
\alias{d_to_r}
\title{Convert Cohen's d to Pearson r}
\usage{
d_to_r(d)
}
\arguments{
\item{d}{Cohen's d}
}
\value{
Pearson r
}
\description{
Convert Cohen's d to Pearson r
}
\examples{
d_to_r(0.5)
}
