% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo.bin.classification}
\alias{demo.bin.classification}
\title{Demonstration of binary GPC}
\usage{
demo.bin.classification()
}
\description{
Run a GPC example.
 Generate class labels using the sigmoid prior GP,
 then learn the posterior GP and predict a new set of data points.
 Plot both samples from the sigmoid prior and the sigmoid posterior
 and the mean prediction of the posterior.
}
\examples{
 demo.bin.classification()
}

