% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{eta2_to_f}
\alias{eta2_to_f}
\title{Convert eta-squared to Cohen's f}
\usage{
eta2_to_f(eta2)
}
\arguments{
\item{eta2}{Eta-squared}
}
\value{
Cohen's f
}
\description{
Convert eta-squared to Cohen's f
}
\examples{
eta2_to_f(0.06)
}
