% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{f2_to_r2}
\alias{f2_to_r2}
\title{Convert Cohen's f-squared to R-squared}
\usage{
f2_to_r2(f2)
}
\arguments{
\item{f2}{Cohen's f-squared}
}
\value{
R-squared
}
\description{
Convert Cohen's f-squared to R-squared
}
\examples{
f2_to_r2(0.15)
}
