% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{f_to_eta2}
\alias{f_to_eta2}
\title{Convert Cohen's f to eta-squared}
\usage{
f_to_eta2(f)
}
\arguments{
\item{f}{Cohen's f}
}
\value{
Eta-squared
}
\description{
Convert Cohen's f to eta-squared
}
\examples{
f_to_eta2(0.25)
}
