% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{f_to_omega2}
\alias{f_to_omega2}
\title{Convert Cohen's f to omega-squared}
\usage{
f_to_omega2(f)
}
\arguments{
\item{f}{Cohen's f}
}
\value{
Omega-squared
}
\description{
Convert Cohen's f to omega-squared
}
\examples{
f_to_omega2(0.25)
}
