% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{gpr-package}
\alias{gpr}
\alias{gpr-package}
\title{gpr package}
\description{
A comprehensive power analysis package for R implementing statistical power analyses similar to those available in G*Power 3.1 (Faul et al., 2007). Supports t-tests, F-tests, chi-square tests, z-tests, and exact tests with programmatic API, interactive console menu, Shiny GUI, and publication-quality plots.
}
\author{
\strong{Maintainer}: Steve Jean-Baptiste \email{sj1021433@wcupa.edu}

}
