% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{interpret_d}
\alias{interpret_d}
\title{Interpret Cohen's d magnitude}
\usage{
interpret_d(d)
}
\arguments{
\item{d}{Cohen's d}
}
\value{
Character string interpretation
}
\description{
Interpret Cohen's d magnitude
}
\keyword{internal}
