% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{interpret_f}
\alias{interpret_f}
\title{Interpret Cohen's f magnitude}
\usage{
interpret_f(f)
}
\arguments{
\item{f}{Cohen's f}
}
\value{
Character string
}
\description{
Interpret Cohen's f magnitude
}
\keyword{internal}
