% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{interpret_f2}
\alias{interpret_f2}
\title{Interpret Cohen's f-squared magnitude}
\usage{
interpret_f2(f2)
}
\arguments{
\item{f2}{Cohen's f-squared}
}
\value{
Character string
}
\description{
Interpret Cohen's f-squared magnitude
}
\keyword{internal}
