% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{interpret_h}
\alias{interpret_h}
\title{Interpret Cohen's h magnitude}
\usage{
interpret_h(h)
}
\arguments{
\item{h}{Cohen's h}
}
\value{
Character string
}
\description{
Interpret Cohen's h magnitude
}
\keyword{internal}
