% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{interpret_w}
\alias{interpret_w}
\title{Interpret Cohen's w magnitude}
\usage{
interpret_w(w)
}
\arguments{
\item{w}{Cohen's w}
}
\value{
Character string
}
\description{
Interpret Cohen's w magnitude
}
\keyword{internal}
