% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{lvgpc.data-class}
\alias{lvgpc.data-class}
\title{Data wrapper for Gausian process classification}
\description{
Data wrapper for Gausian process classification
}
\section{Slots}{

\describe{
\item{\code{x.train}}{the training data for the explanatory variable}

\item{\code{c.train}}{the training data for the labels}

\item{\code{x.new}}{the data for which labels should be predicted}

\item{\code{pars}}{parameter list}
}}

