% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpc.R
\name{lvgpc}
\alias{lvgpc}
\title{Prediction of binomial-distributed random variables using GP classification}
\usage{
lvgpc(x.train = NULL, c.train = NULL, x.new = NULL, pars = list(var = 1,
  inv.scale = 2, gamma = 2, noise = 0.1, kernel = "gamma.exp"), ...)
}
\arguments{
\item{x.train}{vector of independent variables used for training}

\item{c.train}{vector of dependent variables used for training}

\item{x.new}{vector of variables for which a response should be predicted}

\item{pars}{a list containing the hyper-parameters
and kernel specifications}

\item{...}{additional parameters (not specified)}
}
\value{
An object of class \code{lvgpc.pred}

an \code{lvgpc.pred} object
\item{c.predict }{ the predicted c* values given the \code{lvgpc} object}
\item{mean.c.predict }{ the predicted mean c* values
 given the \code{lvgpc} object}
\item{mean }{ the (approximated) posterior mean values}
\item{cov  }{ the (approximated) posterior covariance/kernel}
\item{call  }{ the function call}
}
\description{
Prediction of binomial-distributed
  random variables using GP classification
}
\examples{
 x.train <- seq(-5, 5, .1)
 c.train <- c(rep(0,25), rep(1, length(x.train)-50), rep(0,25))
 x.new <- rnorm(100)
 pars <- list(var=1, inv.scale=2, gamma=2, noise=.1, kernel="gamma.exp")
 pred <- lvgpc(x.train, c.train, x.new, pars)
}
\references{
Rasmussen C.E. and Williams C.K.I. (2006),
 \emph{Gaussian Processes for Machine Learning}, MIT Press \cr
 \url{http://www.gaussianprocess.org/gpml/} \cr \cr
 Barber D. (2013), \emph{Bayesian Reasoning and Machine Learning},
 Cambridge University Press \cr
 \url{http://web4.cs.ucl.ac.uk/staff/D.Barber/pmwiki/pmwiki.php?n=Brml.HomePage}
}

