% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{lvgpr.data-class}
\alias{lvgpr.data-class}
\title{Data wrapper for Gausian process regression}
\description{
Data wrapper for Gausian process regression
}
\section{Slots}{

\describe{
\item{\code{x.train}}{the training data for the explanatory variables}

\item{\code{y.train}}{the training data for the dependant variables}

\item{\code{x.new}}{the data for which the response should be predicted}

\item{\code{pars}}{parameter list}
}}

