% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpr.R
\name{lvgpr}
\alias{lvgpr}
\title{Prediction of normal-distributed random variables using GP regression}
\usage{
lvgpr(x.train = NULL, y.train = NULL, x.new = NULL, pars = list(var = 1,
  inv.scale = 2, gamma = 2, noise = 0.1, kernel = "gamma.exp"), ...)
}
\arguments{
\item{x.train}{vector of independent variables used for training}

\item{y.train}{vector of dependent variables used for training}

\item{x.new}{vector of variables for which a response should be predicted}

\item{pars}{a list containing the hyper-parameters and kernel specifications}

\item{...}{additional parameters (not specified)}
}
\value{
an object of class \code{lvgpr.pred}
\item{y.predict }{ the predicted y* values given the \code{lvgpr} object}
\item{mean }{ the posterior mean values}
\item{cov  }{ the posterior covariance/kernel}
\item{call  }{ the function call}
}
\description{
Prediction of normal-distributed random variables using GP regression
}
\examples{
 x.train <- seq(-5, 5, .1)
 y.train <- rnorm(length(x.train))
 x.new <- rnorm(100)
 pars <- list(var=1, inv.scale=2, gamma=2, noise=.1, kernel="gamma.exp")
 pred <- lvgpr(x.train, y.train, x.new, pars)
}
\references{
Rasmussen C.E. and Williams C.K.I. (2006), \emph{Gaussian Processes for Machine Learning}, MIT Press \cr
 \url{http://www.gaussianprocess.org/gpml/} \cr \cr
 Barber D. (2013), \emph{Bayesian Reasoning and Machine Learning}, Cambridge University Press \cr
 \url{http://web4.cs.ucl.ac.uk/staff/D.Barber/pmwiki/pmwiki.php?n=Brml.HomePage}
}

