% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{mf}
\alias{mf}
\title{Zero mean function}
\usage{
mf(x)
}
\arguments{
\item{x}{Value of independent variable}
}
\value{
Returns 0 for all values
}
\description{
Zero mean function
}
\details{
Zero mean function
}
\author{
Wajid Jawaid
}
