% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{omega2_to_f}
\alias{omega2_to_f}
\title{Convert omega-squared to Cohen's f}
\usage{
omega2_to_f(omega2)
}
\arguments{
\item{omega2}{Omega-squared}
}
\value{
Cohen's f
}
\description{
Convert omega-squared to Cohen's f
}
\examples{
omega2_to_f(0.06)
}
