% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{or_to_d}
\alias{or_to_d}
\title{Convert odds ratio to Cohen's d}
\usage{
or_to_d(or)
}
\arguments{
\item{or}{Odds ratio}
}
\value{
Cohen's d
}
\description{
Convert odds ratio to Cohen's d
}
\examples{
or_to_d(1.68)
}
