% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{or_to_r}
\alias{or_to_r}
\title{Convert odds ratio to Pearson r}
\usage{
or_to_r(or)
}
\arguments{
\item{or}{Odds ratio}
}
\value{
Pearson r
}
\description{
Convert odds ratio to Pearson r
}
\examples{
or_to_r(1.68)
}
