% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{partial_eta2_to_f}
\alias{partial_eta2_to_f}
\title{Convert partial eta-squared to Cohen's f}
\usage{
partial_eta2_to_f(partial_eta2)
}
\arguments{
\item{partial_eta2}{Partial eta-squared}
}
\value{
Cohen's f
}
\description{
Convert partial eta-squared to Cohen's f
}
\examples{
partial_eta2_to_f(0.06)
}
