% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{plot.gpr}
\alias{plot.gpr}
\title{Plot Gaussian process regression object}
\usage{
\method{plot}{gpr}(x, xl = NULL, xu = NULL, yl = NULL, yu = NULL,
  ci = 0.975, ci.col = "#CCCCCC", p.pch = 1, p.cex = 0.5,
  ci.line.col = "darkgray", ci.line.lty = 2, xaxs = "r", xlab = "x",
  ylab = "y", ...)
}
\arguments{
\item{x}{gpr class}

\item{xl}{x-axis lower limit}

\item{xu}{x-axis upper limit}

\item{yl}{y-axis lower limit}

\item{yu}{y-axis upper limit}

\item{ci}{Default 0.975 (5\% CI). Confidence interval to plot}

\item{ci.col}{Default "#CCCCCC". Colour of shaded CI interval}

\item{p.pch}{Default 1. Shape to use for training points}

\item{p.cex}{Default 0.5. Size of training points}

\item{ci.line.col}{Default "darkgray". Colour of confidence interval shaded region outline}

\item{ci.line.lty}{Default 2. Line thickness of confidence interval shaded region outline}

\item{xaxs}{Deafult "i".}

\item{xlab}{Passed to plot}

\item{ylab}{Passed to plot}

\item{...}{Parameters passed to plot e.g. main}
}
\value{
Returns a plot
}
\description{
Plot Gaussian process regression object
}
\details{
Plot Gaussian process regression object
Plots gpr class
}
\author{
Wajid Jawaid
}
