% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{predictGP}
\alias{predictGP}
\title{GP engine - Finds posterior}
\usage{
predictGP(x, y, xs, cvFunc, sigmaf = 1, l = 1, sigman = 0)
}
\arguments{
\item{x}{Measured input values}

\item{y}{Measured output values}

\item{xs}{Where to make predictions}

\item{cvFunc}{Which covariance function to use}

\item{sigmaf}{Passed to \code{\link{cf}}}

\item{l}{Passed to \code{\link{cf}}}

\item{sigman}{Passed to \code{\link{cf}}}
}
\value{
Returns a list containing:
\describe{
  \item{x}{The values where predcitions are made}
  \item{mean}{Posterior mean}
  \item{cv}{Posterior covariance matrix}
  \item{marinal_likelihood}{Marginal likelihood of the model}
}
}
\description{
GP engine - Finds posterior
}
\details{
GP engine - Finds posterior
}
\author{
Wajid Jawaid
}
