% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{r2_to_f2}
\alias{r2_to_f2}
\title{Convert R-squared to Cohen's f-squared}
\usage{
r2_to_f2(r2)
}
\arguments{
\item{r2}{R-squared value}
}
\value{
Cohen's f-squared (f2)
}
\description{
Convert R-squared to Cohen's f-squared
}
\examples{
r2_to_f2(0.13)
}
