% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{r_to_d}
\alias{r_to_d}
\title{Convert Pearson r to Cohen's d}
\usage{
r_to_d(r)
}
\arguments{
\item{r}{Pearson r}
}
\value{
Cohen's d
}
\description{
Convert Pearson r to Cohen's d
}
\examples{
r_to_d(0.243)
}
