% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{r_to_z}
\alias{r_to_z}
\title{Convert Pearson r to Fisher's z}
\usage{
r_to_z(r)
}
\arguments{
\item{r}{Pearson r}
}
\value{
Fisher's z
}
\description{
Convert Pearson r to Fisher's z
}
\examples{
r_to_z(0.35)
}
