% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{sampleFromK}
\alias{sampleFromK}
\title{Sample from Covariance matrix}
\usage{
sampleFromK(L, mu = 0)
}
\arguments{
\item{L}{Cholesky decomposition of Covariance matrix}

\item{mu}{Means to use for sampling}
}
\value{
Vector of predictions
}
\description{
Sample from Covariance matrix
}
\details{
Sample from covariance matrix.
To avoid repeated Cholesky decompositions please supply with lower triangular
Cholesky decomposition of the covariance matrix rather than the covariance
matrix itself.
}
\author{
Wajid Jawaid
}
