% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{sampleFromL}
\alias{sampleFromL}
\title{Sample from Covariance matrix}
\usage{
sampleFromL(L, mu = 0)
}
\arguments{
\item{L}{Cholesky decomposition of Covariance matrix}

\item{mu}{Means to use for sampling}
}
\value{
Vector of predictions
}
\description{
Sample from Covariance matrix
}
\details{
Sample from covariance matrix.
To avoid repeated Cholesky decompositions please supply with lower triangular
Cholesky decomposition of the covariance matrix rather than the covariance
matrix itself.
}
\examples{
set.seed(0)
library(gpr)

d <- .1
x <- seq(-5,5,d)
K <- covMat(x, x, function(xx,xy) cf(xx,xy, l = 2))
L <- cholesky(K)

plot(NULL, xlim=c(-5,5), ylim=c(-2.5,2.5), xlab = "time", ylab = "counts")
rect(-6, -2, 6, 2, col="#44444411", lty=0)
for (i in 1:4) {
  y <- sampleFromL(L)
  points(x,y, pch=16, xlim=c(-5,5), ylim=c(-2,2), cex=.2, col = i, type = "l")
}
}
\author{
Wajid Jawaid
}
