% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{z_to_r}
\alias{z_to_r}
\title{Convert Fisher's z to Pearson r}
\usage{
z_to_r(z)
}
\arguments{
\item{z}{Fisher's z}
}
\value{
Pearson r
}
\description{
Convert Fisher's z to Pearson r
}
\examples{
z_to_r(0.3654)
}
