% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{cholesky}
\alias{cholesky}
\title{Cholesky decomposition}
\usage{
cholesky(K, σ2 = 1e-12, ...)
}
\arguments{
\item{K}{Covariance matrix - symmetric positive definite}

\item{σ2}{Noise term}

\item{...}{Passed to R's built-in function}
}
\value{
Lower triangular matrix
}
\description{
Cholesky decomposition
}
\details{
Cholesky decomposition
The R function returns an upper triangular matrix. Additionally this adds a little noise
to the diagonal
}
\author{
Wajid Jawaid
}
