% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpR_package.R
\docType{package}
\name{gpR-package}
\alias{gpR-package}
\title{gpR}
\description{
\emph{gpR} is a Bayesian machine learning package using latent Gaussian processes.
In general supervised machine learning can be divided in classification,
where we describe data using discrete labels, and regression, where the labels are continuous.
}
\author{
Simon Dirmeier | \email{simon.dirmeier@gmx.de}
}
\keyword{package}

