\name{gpr-package}
\alias{gpr-package}
\alias{gpr}
\docType{package}
\title{
\packageTitle{gpr}
}
\description{
\packageDescription{gpr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gpr}
\packageIndices{gpr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{gpr}

Maintainer: \packageMaintainer{gpr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
