% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{gpr}
\alias{gpr}
\title{Constructor function for gpr S3 class}
\usage{
gpr(pst)
}
\arguments{
\item{pst}{object to be coerced to S3 gpr class}
}
\value{
S3 gpr class object
}
\description{
Constructor function for gpr S3 class
}
\details{
Constructor function for gpr S3 class
}
\author{
Wajid Jawaid
}
