% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.r
\name{predictGP}
\alias{predictGP}
\title{GP engine - Finds posterior}
\usage{
predictGP(x, y, xs, cvFunc, mFunc = mf, sigmaf = 1, l = 1, sigman = 0,
  verbose = FALSE)
}
\arguments{
\item{x}{Measured input values}

\item{y}{Measured output values}

\item{xs}{Where to make predictions}

\item{cvFunc}{Which covariance function to use}

\item{mFunc}{Default 0. Mean function}

\item{sigmaf}{Passed to \code{\link{cf}}}

\item{l}{Passed to \code{\link{cf}}}

\item{sigman}{Passed to \code{\link{cf}}}

\item{verbose}{To print progress}
}
\value{
Returns a list containing:
\describe{
  \item{x}{The values where predcitions are made}
  \item{mean}{Posterior mean}
  \item{cv}{Posterior covariance matrix}
  \item{marinal_likelihood}{Marginal likelihood of the model}
}
}
\description{
GP engine - Finds posterior
}
\details{
GP engine - Finds posterior
}
\examples{
x <- c(-4,-3,-1,0,2)
y <- c(-2,0,1,2,-1)

pst <- predictGP(x, y, xs = seq(-5,5,.1), cvFunc = cf, sigman = 0.1, l = 1)
plot(pst, p.pch = 3, p.cex = 2, xlab = "x", ylab = "y")
}
\author{
Wajid Jawaid
}
